cleaned settings
