+++ apps/web-app/src/app/context/AppShellContexts.tsx
@@ -83,10 +83,6 @@ export interface AppShellCoreContextValue {
   nostrPictureByNpub: Record<string, string | null>;
   paidOverlayIsOpen: boolean;
   paidOverlayTitle: string | null;
-  pendingMintAutoswapChangeConfirmation: {
-    fromMint: string;
-    toMint: string;
-  } | null;
   pendingPaymentMintMeltConfirmation: {
     fromMint: string;
     toMint: string;
@@ -132,14 +128,12 @@ export interface AppShellCoreContextValue {
 
 export interface AppShellActionsContextValue {
   cancelPendingNfcWrite: () => void;
-  closeMintAutoswapChangeConfirmation: () => void;
   closePaymentMintMeltConfirmation: () => void;
   closeMenu: () => void;
   closeShareOptions: () => void;
   closeLnurlWithdrawConfirmation: () => void;
   closeLightningInvoiceConfirmation: () => void;
   closeScan: () => void;
-  confirmMintAutoswapChangeConfirmation: () => void;
   confirmPaymentMintMelt: () => Promise<void>;
   confirmLnurlWithdraw: () => Promise<void>;
   confirmLightningInvoicePayment: () => Promise<void>;
