+++ apps/web-app/src/app/hooks/useAppPreferences.ts
@@ -2,7 +2,6 @@ import React from "react";
 import {
   BANK_PAYMENT_OFFER_RECIPIENT_COUNT_STORAGE_KEY,
   BANK_PAYMENT_OFFER_STAGGER_DELAY_SEC_STORAGE_KEY,
-  CASHU_AUTOSWAP_STORAGE_KEY,
   DECIMAL_AMOUNT_INPUT_STORAGE_KEY,
   DISPLAY_ALLOWED_CURRENCIES_STORAGE_KEY,
   DISPLAY_CURRENCY_STORAGE_KEY,
@@ -16,7 +15,6 @@ import type { DisplayCurrency } from "../../utils/displayAmounts";
 
 interface UseAppPreferencesParams {
   allowedDisplayCurrencies: readonly DisplayCurrency[];
-  cashuAutoswapEnabled: boolean;
   decimalAmountInputEnabled: boolean;
   displayCurrency: DisplayCurrency;
   bankPaymentOfferRecipientCount: number;
@@ -29,7 +27,6 @@ interface UseAppPreferencesParams {
 
 export const useAppPreferences = ({
   allowedDisplayCurrencies,
-  cashuAutoswapEnabled,
   decimalAmountInputEnabled,
   displayCurrency,
   bankPaymentOfferRecipientCount,
@@ -109,17 +106,6 @@ export const useAppPreferences = ({
       // ignore
     }
   }, [bankPaymentOfferStaggerDelaySec]);
-
-  React.useEffect(() => {
-    try {
-      localStorage.setItem(
-        CASHU_AUTOSWAP_STORAGE_KEY,
-        cashuAutoswapEnabled ? "1" : "0",
-      );
-    } catch {
-      // ignore
-    }
-  }, [cashuAutoswapEnabled]);
 
   React.useEffect(() => {
     try {
