+++ apps/web-app/src/i18n/en.ts
@@ -10,11 +10,7 @@ export const en = {
   settingsSecurity: "Security",
   settingsDebug: "Debug",
   nostrInspector: "Inspector",
-  nostrInspectorDescription:
-    "Shows event details, including decrypted message payloads in plain text. Data stays in memory only and is cleared on reload.",
   nostrInspectorLogs: "Collect inspector logs",
-  nostrInspectorLogsDescription:
-    "May include decrypted message payloads in plain text. Stored unencrypted in this device's browser storage for up to 24 hours.",
   nostrInspectorLogsStats: "{count} rows · {size} · oldest row {age} old",
   nostrInspectorLogsLoading: "Loading log statistics…",
   downloadNostrInspectorLogs: "Download logs",
@@ -149,15 +145,11 @@ export const en = {
   payWithCashu: "Pay with cashu",
   preferCashu: "Prefer cashu",
   payWithCashuDisabled: "Pay with cashu is disabled in Advanced settings.",
-  cashuAutoswap: "Auto-swap to main mint",
   showProfileQrOnTilt: "Show profile when phone is turned",
   lightningInvoiceAutoPayLimit: "Auto-pay limit",
-  bankPaymentOfferRecipientCount: "Contacts to share offers with",
   bankPaymentOfferStaggerDelay: "Delay between offer recipients",
   bankPaymentOfferStaggerDelayDecrease: "Decrease delay",
   bankPaymentOfferStaggerDelayIncrease: "Increase delay",
-  bankPaymentOfferRecipientDecrease: "Decrease contact count",
-  bankPaymentOfferRecipientIncrease: "Increase contact count",
   custom: "Custom",
   restore: "Restore",
   claimOwnLightningAddressTitle: "Own lightning address",
@@ -381,15 +373,7 @@ export const en = {
   mintFeeLightningPayments: "Lightning payments",
   setCustomMint: "Set custom mint",
   mintUrlInvalid: "Invalid mint URL.",
-  mintAutoswapChangeWarning:
-    "Linky will swap eligible funds from {fromMint} to the {toMint}.",
-  mintAutoswapChangeWarningTitle: "Funds will be swapped to the new mint",
-  mintAutoswapChangeWarningKeep: "OK",
-  mintAutoswapChangeWarningDisable: "Don't swap anything",
   mintSaved: "Mint saved.",
-  mintSavedAutoswapDisabled: "Mint saved. Auto-swap turned off.",
-  mintSavedAutoswapDisabledTestMint:
-    "Mint saved. Auto-swap turned off for the test mint.",
   mintUpdateFailed: "Mint update failed.",
   mintUpdating: "Updating mint…",
   useDefault: "Use default",
@@ -815,10 +799,7 @@ export const en = {
   unitUsdName: "US dollars",
   unitHiddenName: "Hidden",
   decimalInput: "Decimal places",
-  decimalInputDescription: "Allow cents when entering fiat amounts.",
   sendReadReceipts: "Send read receipts",
-  sendReadReceiptsDescription:
-    "Let contacts see when you have read their messages. You always see their receipts.",
   decimalPoint: ".",
 
   cashuBalance: "Available balance",
