+++ apps/web-app/src/pages/AdvancedPage.tsx
@@ -10,16 +10,12 @@ import {
   Coins,
   Copy,
   Download,
-  FlaskConical,
   Landmark,
   Languages,
   LogOut,
   RadioTower,
-  RefreshCw,
   RotateCw,
   ShieldCheck,
-  Timer,
-  Trash2,
   Upload,
   UserRound,
   Zap,
@@ -30,144 +26,18 @@ import {
 } from "../app/context/AppShellContexts";
 import { useAdvancedSettingsContext } from "../app/context/SystemSettingsContexts";
 import {
-  setInspectorEnabled,
-  setInspectorLogsEnabled,
-  useInspectorEnabled,
-  useInspectorLogsEnabled,
-} from "../devtools/inspector/inspectorEnabled";
-import type { PersistentInspectorLogStats } from "../devtools/inspector/persistentInspectorLogBuffer";
-import {
   countConnectedRelays,
   overallRelayStatus,
   useRelayHealth,
 } from "../app/hooks/useRelayHealth";
-import {
-  LINKY_BANK_PAYMENT_OFFER_MAX_RECIPIENT_COUNT,
-  LINKY_BANK_PAYMENT_OFFER_MAX_STAGGER_DELAY_SEC,
-  LINKY_BANK_PAYMENT_OFFER_MIN_RECIPIENT_COUNT,
-  LINKY_BANK_PAYMENT_OFFER_MIN_STAGGER_DELAY_SEC,
-  LINKY_BANK_PAYMENT_OFFER_STAGGER_DELAY_STEP_SEC,
-} from "../app/lib/bankPaymentOffer";
 import { FeedbackIcon } from "../components/icons";
-import { SettingsStepper } from "../components/SettingsStepper";
+import { SettingsLinkRow, SettingsToggleRow } from "../components/SettingsRows";
 import { useNavigation } from "../hooks/useRouting";
 import { getNativeNotificationPermissionState } from "../platform/nativeBridge";
 import { isNativePlatform } from "../platform/runtime";
 
-interface SettingsLinkRowProps {
-  className?: string;
-  dataGuide?: string;
-  disabled?: boolean;
-  icon: React.ReactNode;
-  label: React.ReactNode;
-  onClick: () => void;
-  tail?: React.ReactNode;
-}
-
-function SettingsLinkRow({
-  className = "",
-  dataGuide,
-  disabled,
-  icon,
-  label,
-  onClick,
-  tail,
-}: SettingsLinkRowProps) {
-  return (
-    <button
-      type="button"
-      className={`settings-row settings-link${className ? ` ${className}` : ""}`}
-      onClick={onClick}
-      disabled={disabled}
-      data-guide={dataGuide}
-    >
-      <span className="settings-left">
-        <span className="settings-icon" aria-hidden="true">
-          {icon}
-        </span>
-        <span className="settings-label">{label}</span>
-      </span>
-      <span className="settings-right">
-        {tail}
-        <span className="settings-chevron" aria-hidden="true">
-          &gt;
-        </span>
-      </span>
-    </button>
-  );
-}
-
-interface SettingsToggleRowProps {
-  checked: boolean;
-  description?: string;
-  disabled?: boolean;
-  icon: React.ReactNode;
-  label: string;
-  onChange: (checked: boolean) => void;
-}
-
-function SettingsToggleRow({
-  checked,
-  description,
-  disabled,
-  icon,
-  label,
-  onChange,
-}: SettingsToggleRowProps) {
-  return (
-    <div className="settings-row">
-      <div className="settings-left">
-        <span className="settings-icon" aria-hidden="true">
-          {icon}
-        </span>
-        {description ? (
-          <span className="settings-label-group settings-label-group-stacked">
-            <span className="settings-label">{label}</span>
-            <span className="settings-description">{description}</span>
-          </span>
-        ) : (
-          <span className="settings-label">{label}</span>
-        )}
-      </div>
-      <label className="switch">
-        <input
-          className="switch-input"
-          type="checkbox"
-          aria-label={label}
-          checked={checked}
-          disabled={disabled}
-          onChange={(event) => onChange(event.target.checked)}
-        />
-      </label>
-    </div>
-  );
-}
-
-const formatInspectorLogSize = (size: number): string => {
-  if (size < 1024) return `${size} B`;
-  if (size < 1024 * 1024) return `${(size / 1024).toFixed(1)} KiB`;
-  return `${(size / (1024 * 1024)).toFixed(1)} MiB`;
-};
-
-const formatInspectorLogAge = (
-  oldestAt: number | null,
-  now: number,
-): string => {
-  if (oldestAt === null) return "—";
-  const ageSeconds = Math.max(0, Math.floor((now - oldestAt) / 1_000));
-  if (ageSeconds < 60) return `${ageSeconds}s`;
-  const ageMinutes = Math.floor(ageSeconds / 60);
-  if (ageMinutes < 60) return `${ageMinutes}m`;
-  return `${Math.floor(ageMinutes / 60)}h`;
-};
-
 export function AdvancedPage(): React.ReactElement {
-  const inspectorEnabled = useInspectorEnabled();
-  const inspectorLogsEnabled = useInspectorLogsEnabled();
   const {
-    bankPaymentOfferRecipientCount,
-    bankPaymentOfferStaggerDelaySec,
-    cashuAutoswapEnabled,
     copyNostrKeys,
     dedupeContacts,
     dedupeContactsIsBusy,
@@ -187,9 +57,6 @@ export function AdvancedPage(): React.ReactElement {
     requestLogout,
     requestPasteNostrKeys,
     seedMnemonic,
-    setBankPaymentOfferRecipientCount,
-    setBankPaymentOfferStaggerDelaySec,
-    setCashuAutoswapEnabled,
     setPayWithCashuEnabled,
   } = useAdvancedSettingsContext();
   const relayHealth = useRelayHealth();
@@ -205,10 +72,6 @@ export function AdvancedPage(): React.ReactElement {
   const { openFeedbackContact, toggleSendReadReceipts } = useAppShellActions();
   const [notificationsEnabled, setNotificationsEnabled] = useState(false);
   const [notificationsIsBusy, setNotificationsIsBusy] = useState(false);
-  const [inspectorLogStats, setInspectorLogStats] =
-    useState<PersistentInspectorLogStats | null>(null);
-  const [inspectorLogActionIsBusy, setInspectorLogActionIsBusy] =
-    useState(false);
   const [armedSecurityAction, setArmedSecurityAction] = useState<
     "copyNostr" | "pasteNostr" | null
   >(null);
@@ -218,16 +81,6 @@ export function AdvancedPage(): React.ReactElement {
   const appVersionLabel = __APP_COMMIT_SHA__
     ? `${__APP_VERSION__} (${__APP_COMMIT_SHA__})`
     : `${__APP_VERSION__}`;
-
-  const inspectorLogStatsLabel = inspectorLogStats
-    ? t("nostrInspectorLogsStats")
-        .replace("{count}", String(inspectorLogStats.rowCount))
-        .replace("{size}", formatInspectorLogSize(inspectorLogStats.totalSize))
-        .replace(
-          "{age}",
-          formatInspectorLogAge(inspectorLogStats.oldestAt, Date.now()),
-        )
-    : t("nostrInspectorLogsLoading");
 
   const getAutoPayLimitLabel = useCallback(
     (limit: number) => {
@@ -288,66 +141,6 @@ export function AdvancedPage(): React.ReactElement {
       clearArmTimeout();
     };
   }, [clearArmTimeout]);
-
-  useEffect(() => {
-    if (!inspectorLogsEnabled) {
-      setInspectorLogStats(null);
-      return;
-    }
-
-    let active = true;
-    let unsubscribe = (): void => undefined;
-    void import("../devtools/inspector/persistentInspectorLogSink")
-      .then(
-        async ({
-          initializePersistentInspectorLogs,
-          subscribePersistentInspectorLogs,
-        }) => {
-          if (!active) return;
-          unsubscribe = subscribePersistentInspectorLogs((stats) => {
-            if (active) setInspectorLogStats(stats);
-          });
-          const stats = await initializePersistentInspectorLogs();
-          if (active) setInspectorLogStats(stats);
-        },
-      )
-      .catch(() => {
-        if (active) pushToast(t("nostrInspectorLogsError"));
-      });
-
-    return () => {
-      active = false;
-      unsubscribe();
-    };
-  }, [inspectorLogsEnabled, pushToast, t]);
-
-  const downloadInspectorLogs = async (): Promise<void> => {
-    setInspectorLogActionIsBusy(true);
-    try {
-      const { downloadPersistentInspectorLogs } =
-        await import("../devtools/inspector/persistentInspectorLogSink");
-      await downloadPersistentInspectorLogs();
-      pushToast(t("nostrInspectorLogsDownloaded"));
-    } catch {
-      pushToast(t("nostrInspectorLogsError"));
-    } finally {
-      setInspectorLogActionIsBusy(false);
-    }
-  };
-
-  const clearInspectorLogs = async (): Promise<void> => {
-    setInspectorLogActionIsBusy(true);
-    try {
-      const { clearPersistentInspectorLogs } =
-        await import("../devtools/inspector/persistentInspectorLogSink");
-      await clearPersistentInspectorLogs();
-      pushToast(t("nostrInspectorLogsCleared"));
-    } catch {
-      pushToast(t("nostrInspectorLogsError"));
-    } finally {
-      setInspectorLogActionIsBusy(false);
-    }
-  };
 
   useEffect(() => {
     let isActive = true;
@@ -491,7 +284,6 @@ export function AdvancedPage(): React.ReactElement {
         <SettingsToggleRow
           icon={<CheckCheck size={18} />}
           label={t("sendReadReceipts")}
-          description={t("sendReadReceiptsDescription")}
           checked={sendReadReceiptsEnabled}
           onChange={toggleSendReadReceipts}
         />
@@ -507,13 +299,6 @@ export function AdvancedPage(): React.ReactElement {
           onChange={setPayWithCashuEnabled}
         />
 
-        <SettingsToggleRow
-          icon={<RefreshCw size={18} />}
-          label={t("cashuAutoswap")}
-          checked={cashuAutoswapEnabled}
-          onChange={setCashuAutoswapEnabled}
-        />
-
         <SettingsLinkRow
           onClick={() => navigateTo({ route: "advancedAutoPayLimit" })}
           icon={<Zap size={18} />}
@@ -524,53 +309,6 @@ export function AdvancedPage(): React.ReactElement {
             </span>
           }
         />
-
-        <div className="settings-row">
-          <div className="settings-left">
-            <span className="settings-icon" aria-hidden="true">
-              <UserRound size={18} />
-            </span>
-            <span className="settings-label">
-              {t("bankPaymentOfferRecipientCount")}
-            </span>
-          </div>
-          <div className="settings-right">
-            <SettingsStepper
-              ariaLabel={t("bankPaymentOfferRecipientCount")}
-              decreaseLabel={t("bankPaymentOfferRecipientDecrease")}
-              increaseLabel={t("bankPaymentOfferRecipientIncrease")}
-              max={LINKY_BANK_PAYMENT_OFFER_MAX_RECIPIENT_COUNT}
-              min={LINKY_BANK_PAYMENT_OFFER_MIN_RECIPIENT_COUNT}
-              onChange={setBankPaymentOfferRecipientCount}
-              step={1}
-              value={bankPaymentOfferRecipientCount}
-            />
-          </div>
-        </div>
-
-        <div className="settings-row">
-          <div className="settings-left">
-            <span className="settings-icon" aria-hidden="true">
-              <Timer size={18} />
-            </span>
-            <span className="settings-label">
-              {t("bankPaymentOfferStaggerDelay")}
-            </span>
-          </div>
-          <div className="settings-right">
-            <SettingsStepper
-              ariaLabel={t("bankPaymentOfferStaggerDelay")}
-              decreaseLabel={t("bankPaymentOfferStaggerDelayDecrease")}
-              increaseLabel={t("bankPaymentOfferStaggerDelayIncrease")}
-              max={LINKY_BANK_PAYMENT_OFFER_MAX_STAGGER_DELAY_SEC}
-              min={LINKY_BANK_PAYMENT_OFFER_MIN_STAGGER_DELAY_SEC}
-              onChange={setBankPaymentOfferStaggerDelaySec}
-              step={LINKY_BANK_PAYMENT_OFFER_STAGGER_DELAY_STEP_SEC}
-              value={bankPaymentOfferStaggerDelaySec}
-              valueText={`${bankPaymentOfferStaggerDelaySec} s`}
-            />
-          </div>
-        </div>
       </div>
 
       <div className="settings-section">
@@ -664,60 +402,10 @@ export function AdvancedPage(): React.ReactElement {
           label={t("reloadApp")}
         />
 
-        <SettingsToggleRow
-          icon={<Bug size={18} />}
-          label={t("nostrInspector")}
-          description={t("nostrInspectorDescription")}
-          checked={inspectorEnabled}
-          onChange={setInspectorEnabled}
-        />
-
-        <SettingsToggleRow
-          icon={<Bug size={18} />}
-          label={t("nostrInspectorLogs")}
-          description={t("nostrInspectorLogsDescription")}
-          checked={inspectorLogsEnabled}
-          onChange={setInspectorLogsEnabled}
-        />
-
-        {inspectorLogsEnabled ? (
-          <div className="inspector-log-stats" aria-live="polite">
-            {inspectorLogStatsLabel}
-          </div>
-        ) : null}
-
-        <SettingsLinkRow
-          onClick={() => void downloadInspectorLogs()}
-          disabled={
-            !inspectorLogsEnabled ||
-            inspectorLogActionIsBusy ||
-            !inspectorLogStats?.rowCount
-          }
-          icon={<Download size={18} />}
-          label={t("downloadNostrInspectorLogs")}
-        />
-
-        <SettingsLinkRow
-          onClick={() => void clearInspectorLogs()}
-          disabled={
-            !inspectorLogsEnabled ||
-            inspectorLogActionIsBusy ||
-            !inspectorLogStats?.rowCount
-          }
-          icon={<Trash2 size={18} />}
-          label={t("clearNostrInspectorLogs")}
-        />
-
         <SettingsLinkRow
           onClick={() => navigateTo({ route: "advancedInspector" })}
           icon={<Bug size={18} />}
-          label={t("openNostrInspector")}
-        />
-
-        <SettingsLinkRow
-          onClick={() => navigateTo({ route: "advancedPushDebug" })}
-          icon={<FlaskConical size={18} />}
-          label="Push / SW Debug (log)"
+          label={t("nostrInspector")}
         />
       </div>
